### 2024-09-10 #7209 @editorialbot
[PRE REVIEW]: DemeterWatch: A Java tool to detect Law of Demeter violations in Java collections
**Submitting author:** @youngkaneda (<a href="http://orcid.org/0009-0004-9334-5936">Juan Pablo Pessoa de Aquino</a>)
**Repository:** https://github.com/youngkaneda/DemeterWatch
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.0
