### 2024-09-10 #7209 @danielskatz
👋 @youngkaneda - thanks for your submission.  As it's not completely clear to me that this is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS, the editors will now discuss this. You should hear back in a week or so.

If you have any comments that will help the editors better understand this, please feel free to add them here.
