### 2024-09-10 #7209 @youngkaneda
> @editorialbot query scope

Hello Daniel 👋. I believe that the DemeterWatch is a fitting example of research software as defined by JOSS criteria. It supports the functioning of research instruments by providing a sophisticated platform for analyzing and understanding complex Java codebases. Specifically, the software aids in the detection of design pattern violations and code smells, which are crucial for maintaining the quality and integrity of software systems used in research environments.

Additionally, DemeterWatch excels in extracting knowledge from large datasets, where the datasets in question are substantial Java code. By employing advanced analysis techniques and generating actionable insights from these projects, the software enables researchers to derive meaningful patterns and anomalies from extensive code collections. We the developers, believe that this functionality is essential for researchers who seek to improve code quality and maintainability in their software systems, thus directly aligning, in our perception, with the JOSS definitions of research software.
