### 2024-10-01 #7209 @vissarion
Hi @msultan14, @louiseadennis, @Kraks would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
