### 2024-10-02 #7209 @msultan14
Hi

Yes I can review it, please can you send a detailed separate email.

Thanks,
Saleem Sultan

On Tue, Oct 1, 2024 at 3:07 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> Hi @msultan14 <https://github.com/msultan14>, @louiseadennis
> <https://github.com/louiseadennis>, @Kraks <https://github.com/Kraks>
> would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
