### 2024-11-04 #7308 @saaikrishnan
@vissarion The paper does not include the `State of the field` section describing how the software compares to other commonly used/existing packages. Everything else looks good to me. Thank you.
