### 2024-11-05 #7308 @vissarion
@saaikrishnan thanks for your review! There is a "Statement of need" section but as you mentioned it does not contain enough comparison and reference to relevant software. Could you please open a related issue in the repository (to discuss it there)?
