### 2024-11-14 youngkaneda/DemeterWatch#1 @saaikrishnan
Missing State of Field section
Can you please include a `State of Field` Section with enough comparison and reference to relevant software ? Thanks
