### 2025-03-05 youngkaneda/DemeterWatch#2 @louiseadennis
Graph is slow to build
I've run demeter-watch on some of my own Java code and succesfully generated calls.txt, elements.json, graph.html and script.js.  When I viewed the graph I got a blank page.  I spent a long time googling how to set up CORS in (one of) my browsers though because I thought this was the issue - based on the README.  I still don't know how to do that and I wonder if it might be useful to include some clear instructions for common browsers.

However, I think the actual issue was that my code generated a large graph, because I looked back at my browser tab and suddenly the graph was there - so I would guess it was just taking a while to build.  It might be worth making a note that for large code bases, the graph may take a while to appear.
