### 2025-03-07 #7308 @louiseadennis
OK - I have three comments which are primarily about documentation.

1. I think the README should include a warning that graphs may be slow to build in the visualiser for large code bases.  I've opened this as an issue at [https://github.com/youngkaneda/DemeterWatch/issues/2](url) .
2. I couldn't spot community guidelines in the README
3. I couldn't find state of the field in teh paper (though that may be because there is nothing comparable).
