### 2025-03-07 #7308 @youngkaneda
@louiseadennis Hello, very appreciate your reports I will work in them ASAP, yeah you are right I will add the warning, the build of the visualization graphs takes a while to load because of the number of nodes in the json, I will also try to improve that, I will add the guidelines also, and about the third point, the other reviewer also said the same, I updated the paper with some information about previous works, but unfortunately there is no other specific projects to relate or compare
