### 2025-03-08 753759d
Update README.md addressing problems pointed by #2
