### 2025-03-08 youngkaneda/DemeterWatch#2 @youngkaneda
@louiseadennis Hello I addressed what was missing in README, about the large codebases increasingly graph loading time I will work on an option to be passed at CLI so the user can choose to build only the graph with the breaks and their path in it, I am right now trying to fix my IntelliJ IDEA setup so I can work on the code, it is not finding the dependencies
