### 2025-03-08 a825b4b
Adding a new option to produce a reduced graph visualization of the method calls, related to #2
