### 2025-03-10 youngkaneda/DemeterWatch#3 @vissarion
JOSS review
Some minor comments on the JOSS manuscript:

- In acknowledgments please specify which author or write "The authors..." if the funding applied to all authors
- It will help the reader if you provide some more references for key notions of the paper e.g. for "Java Collections Framework" and "static analysis"
- Figure 1 has low quality, at least in my setup it looks distorted when I zoom in to read the content, could you please include a higher quality version?
- Could you add the ISBN (I think it is 9780078022128, please check) to ref: "Pressman, R. S., & Maxim, B. R. (2014). Software engineering: A practitioner’s approach (8th ed.). McGraw-Hill Professional" ?
- Ln 79, "c++" should be capitalized (you can use "{}" in bibtex to do this).
-  Law of Demeter is capitalized in the whole text apart from references, please fix it to be consistent in the whole text.
