### 2025-03-10 #7308 @vissarion
@youngkaneda I have filed an issue https://github.com/youngkaneda/DemeterWatch/issues/3 related to minor comments on the manuscript writing.
