### 2025-03-10 f4b2411
adressing paper problems in #3
