+++ paper/paper.bib
@@ -18,7 +18,7 @@
 
 @misc{BLOCK,
     author = {Bock, D.},
-    title = {The Paperboy, The Wallet, and The Law of Demeter},
+    title = {The Paperboy, The Wallet, and The {L}aw of {D}emeter},
     year = {2018},
     url = {https://www2.ccs.neu.edu/research/demeter/demeter-method/LawOfDemeter/paper-boy/demeter.pdf},
     note = {Available online}
@@ -69,34 +69,54 @@
     edition   =  8,
     month     =  jan,
     year      =  2014,
-    address   = "New York, NY"
+    address   = "New York, NY",
+    isbn      = {9780078022128}
 }
 
 @inproceedings{LoDP,
-author = {Chiba, Ryota and Hashiura, Hiroaki and Komiya, Seiichi},
-title = {A tool for detecting detects on class implementation in object oriented program on the basis of the Law of Demeter: focusing on dependency between packages},
-year = {2010},
-isbn = {9789604742318},
-publisher = {World Scientific and Engineering Academy and Society (WSEAS)},
-address = {Stevens Point, Wisconsin, USA},
-abstract = {Design methods, such as OOSE and OMT, are used in object-oriented development. However a common fault of these design methods is that design of classes becomes a problem, therefore we have suggested that lowering the coupling by paying attention to module partition have proposed the method of supporting class design. And extended that the Law of Demeter which is one of the laws to lower coupling, can also be applied to JAVA. In addition, about JAVA, when reusing the modules created by other persons, the application of use is higher with package units than the class units, so made it a rule to apply between packages. We developed the tool, which detected the points that violated this automatically as plug in of Eclipse and inspected it about the effectiveness. It experimented with a made tool, As a result, It was able to detect that violate the Law of Demeter and confirmed that the concrete can be detected that location.},
-booktitle = {Proceedings of the 10th WSEAS International Conference on Applied Computer Science},
-pages = {315–320},
-numpages = {6},
-keywords = {object oriented software development, law of demeter, class design, Eclipse},
-location = {Japan},
-series = {ACS'10}
+    author = {Chiba, Ryota and Hashiura, Hiroaki and Komiya, Seiichi},
+    title = {A tool for detecting detects on class implementation in object oriented program on the basis of the {L}aw of {D}emeter: focusing on dependency between packages},
+    year = {2010},
+    isbn = {9789604742318},
+    publisher = {World Scientific and Engineering Academy and Society (WSEAS)},
+    address = {Stevens Point, Wisconsin, USA},
+    abstract = {Design methods, such as OOSE and OMT, are used in object-oriented development. However a common fault of these design methods is that design of classes becomes a problem, therefore we have suggested that lowering the coupling by paying attention to module partition have proposed the method of supporting class design. And extended that the Law of Demeter which is one of the laws to lower coupling, can also be applied to JAVA. In addition, about JAVA, when reusing the modules created by other persons, the application of use is higher with package units than the class units, so made it a rule to apply between packages. We developed the tool, which detected the points that violated this automatically as plug in of Eclipse and inspected it about the effectiveness. It experimented with a made tool, As a result, It was able to detect that violate the Law of Demeter and confirmed that the concrete can be detected that location.},
+    booktitle = {Proceedings of the 10th WSEAS International Conference on Applied Computer Science},
+    pages = {315–320},
[…]
