### 2025-03-10 youngkaneda/DemeterWatch#3 @youngkaneda
@vissarion I did the fixes needed in the paper, but unfortunately I generated the image again, compared the old with the new after generating the pdf through docker and they were the same, I could read properly even zooming in the paper to 200%. To this matter I do not know how to proceed
