### 2025-03-15 #7308 @danielskatz
(and completely coincidentally, greetings from São Paulo)
