### 2025-03-15 57559d6
minor changes in paper.md
