### 2025-03-15 #7308 @danielskatz
@youngkaneda - I've suggested some small changes in https://github.com/youngkaneda/DemeterWatch/pull/4 - please merge this, or let me know what you disagree with, then we can continue to publication.
