### 2025-03-15 #7308 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
