### 2025-03-15 #7308 @youngkaneda
@danielskatz I've checked the proof, it is fine for me.
